Support macOS guests
Usage:
```
limactl create --video template:macos
limactl start macos
limactl shell macos
```

The password prompt is shown during creating an instance,
so as to run `chown root:wheel ~/.lima/_mnt/0/Library/LaunchDaemons/...`,
which is required for the `lima-macos-init` launch daemon to run.

The password for GUI login is randomly generated and
stored in `/Users/${USER}.guest/password` in the VM.

Fix issue 3618

TODOs (maybe in follow-up PRs):
- [X] Docs
- [X] Switch back the default shell to `zsh`
- [ ] Graceful `limactl stop` (currently it has to be stopped from the guest)
- [ ] Propagate additional cloud-init data
  - [X] Hostname
  - [X] Timezone
  - [X] DNS
  - [ ] CA Certificates
- [ ] Support non-plain mode
  - [ ] mounts
  - [ ] port forwards
- [ ] Let `limactl sudoers` generate the entry for chowning files
  (`chown root:wheel ~/.lima/_mnt/0/Library/LaunchDaemons/...`)
- [ ] Support headless mode (`--video=false`)
- [ ] Support macOS in `hack/update-template.sh`

+++ pkg/iso9660util/iso9660util.go
@@ -19,7 +19,27 @@ type Entry struct {
 	Reader io.Reader
 }
 
-func Write(isoPath, label string, layout []Entry) error {
+type WriteOptions struct {
+	Joliet bool
+}
+
+type WriteOpt func(*WriteOptions)
+
+func WithJoliet() WriteOpt {
+	return func(opts *WriteOptions) {
+		opts.Joliet = true
+	}
+}
+
+func Write(isoPath, label string, layout []Entry, opts ...WriteOpt) error {
+	options := &WriteOptions{}
+	for _, opt := range opts {
+		opt(options)
+	}
+	if options.Joliet {
+		return writeJoliet(isoPath, label, layout)
+	}
+
 	if err := os.RemoveAll(isoPath); err != nil {
 		return err
 	}
